refactor: 사용하지 않는 dto 코드 제거(member.dto.ts)
